+++ common/src/main/java/net/satisfy/candlelight/core/block/entity/TableSetBlockEntity.java
@@ -4,10 +4,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
-import net.satisfy.candlelight.core.registry.EntityTypeRegistry;
-import org.jetbrains.annotations.NotNull;
 
 public class TableSetBlockEntity extends StorageBlockEntity {
     private ItemStack effectStack = ItemStack.EMPTY;
@@ -19,16 +16,6 @@ public class TableSetBlockEntity extends StorageBlockEntity {
 
     public TableSetBlockEntity(BlockPos pos, BlockState state, int size) {
         super(pos, state, size);
-    }
-
-    @Override
-    public @NotNull BlockEntityType<?> getType() {
-        return EntityTypeRegistry.TABLE_SET_BLOCK_ENTITY.get();
-    }
-
-    @Override
-    public boolean isValidBlockState(BlockState blockState) {
-        return getType().isValid(blockState);
     }
 
     public ItemStack getEffectStack() {
@@ -48,22 +35,21 @@ public class TableSetBlockEntity extends StorageBlockEntity {
     @Override
     protected void loadAdditional(CompoundTag compoundTag, HolderLookup.Provider provider) {
         super.loadAdditional(compoundTag, provider);
-        if (!this.effectStack.isEmpty()) {
-            if (this.level != null) {
-                this.effectStack = ItemStack.parseOptional(this.level.registryAccess(), compoundTag.getCompound("EffectStack"));
-                this.effectDuration = compoundTag.getInt("EffectDuration");
-            }
+        if (this.level != null && compoundTag.contains("EffectStack")) {
+            this.effectStack = ItemStack.parseOptional(this.level.registryAccess(), compoundTag.getCompound("EffectStack"));
+            this.effectDuration = compoundTag.getInt("EffectDuration");
+        } else {
+            this.effectStack = ItemStack.EMPTY;
+            this.effectDuration = 0;
         }
     }
 
     @Override
     protected void saveAdditional(CompoundTag nbt, HolderLookup.Provider provider) {
         super.saveAdditional(nbt, provider);
-        if (!this.effectStack.isEmpty()) {
-            if (this.level != null) {
-                nbt.put("EffectStack", this.effectStack.save(this.level.registryAccess(), new CompoundTag()));
-                nbt.putInt("EffectDuration", this.effectDuration);
-            }
+        if (this.level != null && !this.effectStack.isEmpty()) {
+            nbt.put("EffectStack", this.effectStack.save(this.level.registryAccess(), new CompoundTag()));
+            nbt.putInt("EffectDuration", this.effectDuration);
         }
     }
 }